+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Route_OM/OM_Plan_A.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Plan_For_OM;
+package org.firstinspires.ftc.teamcode.Route_OM;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
@@ -6,11 +6,13 @@ import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import org.firstinspires.ftc.teamcode.Libs.Core_Base_Platform;
 
 @Disabled
-@Autonomous(name = "OM Plan B",group = "OM Contest")
-public final class OM_Plan_B extends Core_Base_Platform {
+@Autonomous(name = "OM Plan A",group = "OM Contest")
+
+public final class OM_Plan_A extends Core_Base_Platform {
 
     @Override
     public void runOpMode(){
+        initRobot();
 
     }
 }
